feat : 최근 상품 조회 api 입니다.

+++ core/src/main/kotlin/application/port/in/use_case/AlilmHistoryUseCase.kt
@@ -6,6 +6,6 @@ interface AlilmHistoryUseCase {
 
     data class AlilmHistoryResult(
         val allCount : Long,
-        val dailyCount : Long,
+        val dailyCount : Int,
     )
 }
